checkers/requirements.txt
checklib==0.7.0
requests==2.31.0
beautifulsoup4==4.12.2
checklib==0.5.2
pwntools==4.10.0
pycryptodome==3.18.0
pyelftools==0.29